requirements.txt
asgiref==3.5.2
Django==3.2.15
djangorestframework==3.13.1
pycryptodome==3.15.0
pytz==2022.2.1
sqlparse==0.4.2
